moved announcement component into annoucnements and moved styling for authmodal to component folder
