+++ weddingbells/src/components/Announcements/Announcement.js
